+++ src/components/Landing/index.jsx
@@ -3,7 +3,7 @@ import { node, string } from 'prop-types';
 import classNames from 'classnames';
 import { withStyles } from '@material-ui/core/styles';
 import PageTitle from '../PageTitle';
-import ErrorPanel from '../ErrorPanel/errorPanel';
+import ErrorPanel from '../ErrorPanel';
 
 @withStyles(theme => ({
   root: {
