+++ src/views/AwsProvisioner/ViewWorkerTypeDefinition/index.jsx
@@ -22,7 +22,7 @@ import removeKeys from '../../../utils/removeKeys';
 import formatError from '../../../utils/formatError';
 import isWorkerTypeNameValid from '../../../utils/isWorkerTypeNameValid';
 import { DEFAULT_AWS_WORKER_TYPE } from '../../../utils/constants';
-import ErrorPanel from '../../../components/ErrorPanel/errorPanel';
+import ErrorPanel from '../../../components/ErrorPanel';
 import workerTypeQuery from './workerType.graphql';
 import updateAwsProvisionerWorkerTypeQuery from './updateAwsProvisionerWorkerType.graphql';
 import createAwsProvisionerWorkerTypeQuery from './createAwsProvisionerWorkerType.graphql';
