+++ src/views/Clients/ViewClients/index.jsx
@@ -13,7 +13,7 @@ import ClientsTable from '../../../components/ClientsTable';
 import { VIEW_CLIENTS_PAGE_SIZE } from '../../../utils/constants';
 import clientsQuery from './clients.graphql';
 import { withAuth } from '../../../utils/Auth';
-import ErrorPanel from '../../../components/ErrorPanel/errorPanel';
+import ErrorPanel from '../../../components/ErrorPanel';
 
 @hot(module)
 @withAuth
