+++ src/views/Display/index.jsx
@@ -9,7 +9,7 @@ import TableCellListItem from '../../components/TableCellListItem';
 import Dashboard from '../../components/Dashboard';
 import DataTable from '../../components/DataTable';
 import { VNC_DISPLAYS_POLLING_INTERVAL } from '../../utils/constants';
-import ErrorPanel from '../../components/ErrorPanel/errorPanel';
+import ErrorPanel from '../../components/ErrorPanel';
 
 @withStyles({
   vncDisplay: {
